+++ stormpath/resource/__init__.py
@@ -1,8 +1,12 @@
+"""All Stormpath API resources."""
+
+
 from .account import Account, AccountList
 from .group import Group, GroupList
 from .tenant import Tenant
 from .directory import Directory
 from .group_membership import GroupMembership, GroupMembershipList
+from .custom_data import CustomData
 from .base import (Expansion, Resource, CollectionResource,
-    SaveMixin, DeleteMixin)
+    SaveMixin, DeleteMixin, AutoSaveMixin)
 from .password_reset_token import PasswordResetTokenList