+++ stormpath/resource/account_store.py
@@ -1,9 +1,11 @@
+"""Stormpath AccountStore resource mappings."""
+
+
 def AccountStore(client, properties=None):
-    """Account store resource factory
+    """AccountStore resource factory.
 
     Returns either a Group or a Directory resource, based on the resource
     href.
-
     """
     from .directory import Directory
     from .group import Group