+++ tests/httprettys/test_group.py
@@ -11,22 +11,26 @@ class TestGroup(BaseTest):
 
     @httpretty.activate
     def test_properties(self):
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.base_href + "/tenants/current",
             location=self.tenant_href,
             status=302)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.dir_href,
             body=json.dumps(self.dir_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.grp_href,
             body=json.dumps(self.grp_body),
             content_type="application/json")
@@ -42,26 +46,31 @@ class TestGroup(BaseTest):
 
     @httpretty.activate
     def test_delete(self):
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.base_href + "/tenants/current",
             location=self.tenant_href,
             status=302)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.dir_href,
             body=json.dumps(self.dir_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.DELETE,
+        httpretty.register_uri(
+            httpretty.DELETE,
             self.grp_href,
             body='', status=204)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.grp_href,
             body=json.dumps(self.grp_body),
             content_type="application/json")
@@ -71,27 +80,30 @@ class TestGroup(BaseTest):
         group.delete()
 
         self.assertEqual(HTTPretty.last_request.method, 'DELETE')
-        self.assertEqual(HTTPretty.last_request.path,
-            self.grp_path)
+        self.assertEqual(HTTPretty.last_request.path, self.grp_path)
 
     @httpretty.activate
     def test_update(self):
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.base_href + "/tenants/current",
             location=self.tenant_href,
             status=302)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.dir_href,
             body=json.dumps(self.dir_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.grp_href,
             body=json.dumps(self.grp_body),
             content_type="application/json")
@@ -119,7 +131,8 @@ class TestGroup(BaseTest):
             }
         }
 
-        httpretty.register_uri(httpretty.POST,
+        httpretty.register_uri(
+            httpretty.POST,
             self.grp_href,
             body=json.dumps(self.updated_grp_body),
             content_type="application/json")
@@ -133,8 +146,7 @@ class TestGroup(BaseTest):
         group.save()
 
         self.assertEqual(HTTPretty.last_request.method, 'POST')
-        self.assertEqual(HTTPretty.last_request.path,
-            self.grp_path)
+        self.assertEqual(HTTPretty.last_request.path, self.grp_path)
 
         self.assertEqual(group.name, name)
         self.assertEqual(group.description, desc)
@@ -142,22 +154,26 @@ class TestGroup(BaseTest):
 
     @httpretty.activate
     def test_create(self):
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.base_href + "/tenants/current",
             location=self.tenant_href,
             status=302)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.dir_href,
             body=json.dumps(self.dir_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.POST,
+        httpretty.register_uri(
+            httpretty.POST,
             self.dir_href + "/groups",
             body=json.dumps(self.grp_body),
             status=201,
@@ -167,37 +183,44 @@ class TestGroup(BaseTest):
         directory.groups.create(self.grp_body)
 
         self.assertEqual(HTTPretty.last_request.method, 'POST')
-        self.assertEqual(HTTPretty.last_request.path,
+        self.assertEqual(
+            HTTPretty.last_request.path,
             self.dir_path + "/groups")
 
     @httpretty.activate
     def test_add_account(self):
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.base_href + "/tenants/current",
             location=self.tenant_href,
             status=302)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.dir_href,
             body=json.dumps(self.dir_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.grp_href,
             body=json.dumps(self.grp_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.acc_href,
             body=json.dumps(self.acc_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.POST,
+        httpretty.register_uri(
+            httpretty.POST,
             self.dir_href + '/groupMemberships',
             body=json.dumps(self.grp_body),
             content_type="application/json")
@@ -212,7 +235,8 @@ class TestGroup(BaseTest):
             "account": {"href": account.href},
             "group": {"href": group.href},
         }
-        httpretty.register_uri(httpretty.POST,
+        httpretty.register_uri(
+            httpretty.POST,
             grp_mem_href,
             body=json.dumps(grp_mem_body),
             content_type="application/json")
@@ -224,86 +248,93 @@ class TestGroup(BaseTest):
 
         self.assertIsInstance(group_membership, GroupMembership)
         self.assertTrue(group_membership.href)
-        self.assertEqual(group_membership.account.email, self.acc_body['email'])
+        self.assertEqual(
+            group_membership.account.email, self.acc_body['email'])
         self.assertEqual(group_membership.group.name, self.grp_body['name'])
 
-    # FIXME: groups are not implemented on server side
-    # although it is in rest api docs
-    # @httpretty.activate
-    # def test_search(self):
-    #     # fetch group
-    #     httpretty.register_uri(httpretty.GET,
-    #         self.base_url + "/tenants/current",
-    #         body=json.dumps(self.tenant_body),
-    #         content_type="application/json")
+    @httpretty.activate
+    def test_search(self):
+        # fetch group
+        httpretty.register_uri(
+            httpretty.GET,
+            self.base_url + "/tenants/current",
+            body=json.dumps(self.tenant_body),
+            content_type="application/json")
 
-    #     httpretty.register_uri(httpretty.GET,
-    #         self.grp_href,
-    #         body=json.dumps(self.grp_body),
-    #         content_type="application/json")
+        httpretty.register_uri(
+            httpretty.GET,
+            self.grp_href,
+            body=json.dumps(self.grp_body),
+            content_type="application/json")
 
-    #     group = self.client.groups.get(self.grp_href)
+        group = self.client.groups.get(self.grp_href)
 
-    #     self.resource_body = {
-    #         "href": self.grp_href + "/accounts",
-    #         "offset": 0,
-    #         "limit": 25,
-    #         "items": [self.acc_href]
-    #     }
+        self.resource_body = {
+            "href": self.grp_href + "/accounts",
+            "offset": 0,
+            "limit": 25,
+            "items": [self.acc_body]
+        }
 
-    #     # search group accounts without attributes
-    #     httpretty.register_uri(httpretty.GET,
-    #         self.grp_href + "/accounts",
-    #         body=json.dumps(self.resource_body),
-    #         content_type="application/json")
+        # search group accounts without attributes
+        httpretty.register_uri(
+            httpretty.GET,
+            self.grp_href + "/accounts",
+            body=json.dumps(self.resource_body),
+            content_type="application/json")
 
-    #     accounts = group.accounts.search("ACC_USERNAME")
-    #     for acc in accounts:
-    #         pass
+        accounts = group.accounts.search("ACC_USERNAME")
+        for acc in accounts:
+            pass
 
-    #     self.assertIsInstance(accounts, AccountResourceList)
-    #     self.assertEqual(HTTPretty.last_request.method, 'GET')
-    #     self.assertEqual(HTTPretty.last_request.path,
-    #         "%s/accounts?q=%s" % (self.app_path,
-    #                                     self.acc_body['username']))
+        self.assertIsInstance(accounts, AccountList)
+        self.assertEqual(HTTPretty.last_request.method, 'GET')
+        self.assertEqual(
+            HTTPretty.last_request.path,
+            "%s/accounts?q=%s" % (self.grp_path, self.acc_body['username']))
 
-    #     # search group accounts with attributes
-    #     httpretty.register_uri(httpretty.GET,
-    #         self.grp_href + "/accounts?givenName=ACC_USERNAME",
-    #         body=json.dumps(self.resource_body),
-    #         content_type="application/json")
+        # search group accounts with attributes
+        httpretty.register_uri(
+            httpretty.GET,
+            self.grp_href + "/accounts?givenName=ACC_USERNAME",
+            body=json.dumps(self.resource_body),
+            content_type="application/json")
 
-    #     accounts = group.accounts.search({
-    #        "given_name": "ACC_USERNAME"})
-    #     for acc in accounts:
-    #         pass
+        accounts = group.accounts.search({"given_name": "ACC_USERNAME"})
+        for acc in accounts:
+            pass
 
-    #     self.assertIsInstance(accounts, AccountResourceList)
-    #     self.assertEqual(HTTPretty.last_request.method, 'GET')
-    #     self.assertEqual(HTTPretty.last_request.path,
-    #        "%s/accounts?givenName=%s" % (self.grp_path,
-    #                                    self.acc_body['username']))
+        self.assertIsInstance(accounts, AccountList)
+        self.assertEqual(HTTPretty.last_request.method, 'GET')
+        self.assertEqual(
+            HTTPretty.last_request.path,
+            "%s/accounts?givenName=%s" % (
+                self.grp_path, self.acc_body['username']))
 
     @httpretty.activate
     def test_associations(self):
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.base_href + "/tenants/current",
             location=self.tenant_href,
             status=302)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.dir_href,
             body=json.dumps(self.dir_body),
             content_type="application/json")
 
         directory = self.client.directories.get(self.dir_href)
 
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.grp_href,
             body=json.dumps(self.grp_body),
             content_type="application/json")
@@ -311,7 +342,8 @@ class TestGroup(BaseTest):
         group = directory.groups.get(self.grp_href)
 
         # accounts
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.acc_href,
             body=json.dumps(self.acc_body),
             content_type="application/json")
@@ -323,7 +355,8 @@ class TestGroup(BaseTest):
         self.assertEqual(group.directory.href, self.dir_href)
 
         # tenant
-        httpretty.register_uri(httpretty.GET,
+        httpretty.register_uri(
+            httpretty.GET,
             self.tenant_href,
             body=json.dumps(self.tenant_body),
             content_type="application/json")